ai-agent-formatting/pyproject.toml
[project]
name = "ai-agent-formatting"
version = "0.0.1"
description = ""
authors = [
    {name = "Thiago Rufino",email = "thiago.rufino@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "black (>=25.1.0,<26.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "langchain-openai (>=0.3.22,<0.4.0)",
    "load-dotenv (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
